refactor handlePaymentGatewayError into object
